requirements.txt
# Core dependencies
neo4j>=5.15.0
rich>=13.7.0
click>=8.1.7
pydantic>=2.5.2
autogen-core>=0.4.9
loguru>=0.7.2
typer>=0.9.0
protobuf>=4.24.4
grpcio>=1.59.2
grpcio-tools>=1.59.2
blarify>=0.1.0
prompty>=0.1.0

# Azure and OpenAI dependencies
azure-identity>=1.15.0
azure-mgmt-cognitiveservices>=13.5.0
openai>=1.2.4

# GitHub integration
pygithub>=2.1.1
gitpython>=3.1.44
tqdm>=4.67.1

# Development dependencies
black>=23.11.0
mypy>=1.7.0
isort>=5.12.0
flake8>=6.1.0
pre-commit>=3.5.0
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
types-protobuf>=4.24.0.2
pydocstyle>=6.3.0
docstr-coverage>=2.3.0
bandit>=1.7.5
poethepoet>=0.33.1

# Documentation dependencies
sphinx>=7.2.6
sphinx-autoapi>=2.1.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.25.2